microservices/serving/recommendation_loader/requirements.txt
pyspark>=3.4,<4.0
pyyaml>=6.0
psycopg2-binary>=2.9
